kube-proxy: windows: Removed hnsV1
        hnsV1 not supported anymore
